requirements.txt
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pydantic==2.11.3
pydantic-settings==2.10.1
pydantic_core==2.33.1
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
python-jose==3.5.0
psycopg2-binary==2.9.10
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
